+++ flink-ml-core/pom.xml
@@ -34,6 +34,13 @@ under the License.
   <dependencies>
     <dependency>
       <groupId>org.apache.flink</groupId>
+      <artifactId>flink-ml-servable-core</artifactId>
+      <version>${project.version}</version>
+      <scope>provided</scope>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.flink</groupId>
       <artifactId>flink-ml-iteration</artifactId>
       <version>${project.version}</version>
       <scope>provided</scope>
