+++ flink-ml-servable-core/src/main/java/org/apache/flink/ml/linalg/VectorWithNorm.java
