+++ flink-ml-servable-core/src/main/java/org/apache/flink/ml/param/VectorParam.java
