+++ flink-ml-uber/pom.xml
@@ -55,6 +55,18 @@ under the License.
 
     <dependency>
       <groupId>org.apache.flink</groupId>
+      <artifactId>flink-ml-servable-core</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.flink</groupId>
+      <artifactId>flink-ml-servable-lib</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.flink</groupId>
       <artifactId>flink-ml-benchmark</artifactId>
       <version>${project.version}</version>
     </dependency>
@@ -78,6 +90,8 @@ under the License.
                   <include>org.apache.flink:flink-ml-core</include>
                   <include>org.apache.flink:flink-ml-iteration</include>
                   <include>org.apache.flink:flink-ml-lib</include>
+                  <include>org.apache.flink:flink-ml-servable-core</include>
+                  <include>org.apache.flink:flink-ml-servable-lib</include>
                   <include>org.apache.flink:flink-ml-benchmark</include>
                   <include>dev.ludovic.netlib:blas</include>
                 </includes>
